requirements.txt
# requirements.txt

# UI
streamlit==1.36.0

# OpenAI SDK
openai==1.102.0

# Extra dependencies
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0